split out build provider tests
